# skip boilerplate check for minikube-image-benchmark submodule